% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{BuildBoundaryPoly}
\alias{BuildBoundaryPoly}
\title{Build boundary polygon based on the boundary points}
\usage{
BuildBoundaryPoly(boundary = NULL)
}
\arguments{
\item{boundary}{Boundary result returned by \code{GetBoundary} function.}
}
\description{
Build boundary polygon based on the boundary points
}
