% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeWeights.R
\name{ComputeCentroidWeights}
\alias{ComputeCentroidWeights}
\title{Compute spatial weights based on distance to centroid}
\usage{
ComputeCentroidWeights(
  data = NULL,
  cell_ids = NULL,
  scale = TRUE,
  method = c("inverse", "gaussian", "linear", "quadratic"),
  sigma = 0.5
)
}
\arguments{
\item{data}{A data frame or Seurat object containng at least three columns: \code{x}, \code{y} and \code{cell}.}

\item{cell_ids}{A vector of cell IDs to subset from data.}

\item{scale}{A logical value indicating whether to scale distances to [0, 1] before computing weights. Scaling is recommended, and default is TRUE.}

\item{method}{A string specifying the decay method. One of \code{"inverse"},
\code{"gaussian"}, \code{"linear"}, or \code{"quadratic"}.}

\item{sigma}{Standard deviation for the Gaussian decay function. Default is 0.5 for scaled distance. If \code{scale = FALSE}, choose a value that is about half the region’s width.}
}
\value{
A named numeric vector of weights, one per cell in \code{cell_ids}.
}
\description{
Computes spatial weights for a set of cells based on their distance to
the centroid of their group. Supports multiple decay methods and optional
distance scaling.
}
