% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialEnrichmentIndex.R
\name{computeSEI}
\alias{computeSEI}
\title{Compute Spatial Enrichment Index (SEI)}
\usage{
computeSEI(exp_mat = NULL, weights = NULL)
}
\arguments{
\item{exp_mat}{A normalized gene expression matrix with genes as rows and
cells as columns. Should be of class \code{matrix} or \code{dgCMatrix}.}

\item{weights}{A numeric vector of spatial weights (e.g.,
from \code{computeBoundaryWeights} or \code{computeCentroidWeights}).
Must be the same length as the number of columns (cells)
in \code{exp_mat}.}
}
\value{
A data frame containing the spatial enrichment index (SEI) results.
}
\description{
Alias for \code{\link{computeSpatialEnrichmentIndex}}.
}
\examples{
# Load spatial coordinates and log-normalized expression
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
    package = "SpNeigh"
))
logNorm_expr <- readRDS(system.file("extdata", "LogNormExpr.rds",
    package = "SpNeigh"
))

# Compute spatial weights and SEI
bon_c0 <- getBoundary(data = coords, one_cluster = 0)
cells_c0 <- subset(coords, cluster == 0)$cell
weights <- computeBoundaryWeights(
    data = coords,
    cell_ids = cells_c0,
    boundary = bon_c0
)
sei_df <- computeSEI(logNorm_expr[, cells_c0], weights)
head(sei_df)
}
