% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellsInside.R
\name{GetCellsInside}
\alias{GetCellsInside}
\title{Get cells inside boundaries or rings}
\usage{
GetCellsInside(data = NULL, boundary = NULL)
}
\arguments{
\item{data}{Either a Seurat object or a data frame with columns: \code{x}, \code{y}, \code{cell} and \code{cluster}.}

\item{boundary}{A sf object or a data frame. Boundary result returned by \code{GetBoundary} or \code{GetBoundaryPoly} functions. Or ring regions returned by \code{GetRingRegion}.}
}
\description{
Get cells inside boundaries or rings
}
\examples{
# Load coordinates
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
                              package = "SpNeigh"))
head(coords)

# Get cells inside boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
cells_inside <- GetCellsInside(data = coords, boundary =  boundary_points)
cells_inside

# Get cells inside rings
ring_regions <- GetRingRegion(boundary = boundary_points, dist = 100)
cells_ring <- GetCellsInside(data = coords, boundary =  ring_regions)
cells_ring

}
