% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{GetOuterBoundary}
\alias{GetOuterBoundary}
\title{Get outer boundary of the original boundary}
\usage{
GetOuterBoundary(boundary = NULL, dist = 100)
}
\arguments{
\item{boundary}{A data frame or sf object. Boundary result returned by \code{GetBoundary} or \code{GetBoundaryPoly} function.}

\item{dist}{Distance of the outer boundary from the existing boundary.}
}
\description{
Get outer boundary of the original boundary
}
