% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{GetRingRegion}
\alias{GetRingRegion}
\title{Get ring region between the original boundary and outer boundary}
\usage{
GetRingRegion(boundary = NULL, outer_boundary = NULL, ...)
}
\arguments{
\item{boundary}{A data frame or sf object. Boundary result returned by \code{GetBoundary} or \code{GetBoundaryPoly} function.}

\item{outer_boundary}{A sf object. Outer boundary returned by \code{GetOuterBoundary} function.}

\item{...}{Parameters in \code{GetOuterBoundary}}
}
\description{
Get ring region between the original boundary and outer boundary
}
