% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_colors_15}
\alias{my_colors_15}
\title{Custom Color Palette}
\format{
A character vector of length 15.
}
\usage{
my_colors_15
}
\description{
A character vector containing 15 colors.
}
\examples{
plot(1:15, 1:15, pch = 16, col = my_colors_15)
barplot(rep(1, 15), col = my_colors_15)

}
\keyword{datasets}
