% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{my_theme_ggplot}
\alias{my_theme_ggplot}
\title{A Custom ggplot2 Theme}
\usage{
my_theme_ggplot()
}
\value{
A ggplot2 theme object.
}
\description{
This is a ggplot2 theme defined and used in this package.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  my_theme_ggplot()
}
