% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{PlotBoundary}
\alias{PlotBoundary}
\title{Draw boundary for a cluster or cell population}
\usage{
PlotBoundary(
  data = NULL,
  one_cluster = NULL,
  boundary = NULL,
  colors = my_colors_15,
  point_size = 0.5,
  color_boundary = "black",
  linewidth_boundary = 1.5,
  sub_plot = FALSE,
  theme_ggplot = my_theme_ggplot(),
  legend_size = 2,
  ...
)
}
\arguments{
\item{data}{A data frame contains coordinates of the cells and cluster information. If the input is a Seurat object, the information will be extracted automatically.}

\item{one_cluster}{Specify one cluster to obtain its boundary. If \code{boundary} parameter is provided, it can be omitted. However, it must be provided, if \code{sub_plot} is TRUE.}

\item{boundary}{Boundary obtained from \code{GetBoundary} function. If it is NULL, get the boundary of the specified cluster automatically. If both \code{boundary} and \code{one_cluster} are NULL, plot coordinates only.}

\item{colors}{A vector of colors used in plotting for clusters.}

\item{point_size}{Point size of cells. Default is 0.5.}

\item{color_boundary}{Color of the boundary for a cluster. Default is black.}

\item{linewidth_boundary}{Linewidth of the boundary for a cluster. Default is 1.5.}

\item{sub_plot}{A logical value indicates whether to only plot cells in one cluster. If FALSE, plot all cells in all clusters. Default is FALSE.}

\item{theme_ggplot}{Theme for ggplot.}

\item{legend_size}{Legend size for ggplot.}

\item{...}{Parameters in \code{GetBoundary} function.}
}
\description{
Draw boundary for a cluster or cell population
}
