% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{PlotEdge}
\alias{PlotEdge}
\title{Plot boundary edge}
\usage{
PlotEdge(
  boundary_poly = NULL,
  linewidth_boundary = 1,
  theme_ggplot = my_theme_ggplot(),
  ...
)
}
\arguments{
\item{boundary_poly}{An sf object containing x, y and region_id information.}

\item{linewidth_boundary}{Linewidth of the boundaries. Default is 1.}

\item{theme_ggplot}{Theme for ggplot.}

\item{...}{Other parameters in \code{sf::geom_sf()} function.}
}
\description{
Plot boundary edge
}
\examples{
# Load coordinates
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
                              package = "SpNeigh"))
head(coords)

# Plot region inside boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
boundary_polys = BuildBoundaryPoly(boundary_points)
PlotEdge(boundary_poly = boundary_polys)

# Split boundary polygon 1 into two edges
boundary_edges <- SplitBoundaryPolyByAnchor(boundary_polys[1,])
PlotEdge(boundary_edges)

}
