% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeights.R
\name{PlotWeights}
\alias{PlotWeights}
\title{Plot spatial weights for cells}
\usage{
PlotWeights(
  data = NULL,
  weights = NULL,
  point_size = 0.2,
  theme_ggplot = my_theme_ggplot()
)
}
\arguments{
\item{data}{A data frame or Seurat object containng at least three columns: \code{x}, \code{y} and \code{cell}.}

\item{weights}{A named numeric vector where names are cell IDs.}

\item{point_size}{Point size in the plot. Default is 0.2.}

\item{theme_ggplot}{Theme for ggplot.}
}
\description{
Plots spatial coordinates of cells colored by weight values. Only cells in the \code{data} with weights will be plotted.
}
\examples{
# Load coordinates
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
                              package = "SpNeigh"))
head(coords)

# Obtain cells in cluster 2 and focus on these cells
cells_c2 <- subset(coords, cluster==2)[,"cell"]

# Obtain and plot centroid weights
weights_cen <- ComputeCentroidWeights(data = coords, cell_ids = cells_c2)
PlotWeights(data = coords, weights = weights_cen)

# Obtain and plot boundary weights
boundary_points <- GetBoundary(data = coords, one_cluster = 2)
weights_bon <- ComputeBoundaryWeights(data = coords, cell_ids = cells_c2,
                                      boundary = boundary_points)
PlotWeights(data = coords, weights = weights_bon)

}
