% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDE.R
\name{RunSpatialDE}
\alias{RunSpatialDE}
\title{Differential expression analysis along spatial distance gradients}
\usage{
RunSpatialDE(
  exp_mat = NULL,
  cells_ids = NULL,
  spatial_distance = NULL,
  adj_p.value = 0.05
)
}
\arguments{
\item{exp_mat}{A normalized gene expression matrix with genes as rows and cells as columns. For example, the log-normalized data in a Seurat object.}

\item{cells_ids}{A vector of cell IDs in \code{exp_mat} for spatial differential expression analysis.}

\item{spatial_distance}{A numeric vector of the spatial distance for each cell. Scaled distance rather than real distance is is recommended.}

\item{adj_p.value}{Adjusted p-value threshold for significance to obtain differentially expressed genes. Default is 0.05.}
}
\value{
A data.frame of differentially expressed genes along spatial distance. Positive trend means positive correlation of the gene expression with the spatial distance (Up regulated), and negative trend means negative correlation (Down regulated).
}
\description{
Differential expression analysis along spatial distance gradients
}
