% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safeColorPalette}
\alias{safeColorPalette}
\title{Generate a safe color palette for discrete clusters}
\usage{
safeColorPalette(
  n_clusters = NULL,
  base_colors = colors15_cheng,
  verbose = TRUE
)
}
\arguments{
\item{n_clusters}{Number of unique clusters.}

\item{base_colors}{A character vector of base colors.
Default is \code{colors15_cheng}.}

\item{verbose}{Logical. If \code{TRUE}, displays a message when
extended colors are used. Default is \code{TRUE}.}
}
\value{
A character vector of colors of length \code{n_clusters}.
}
\description{
Returns a vector of visually distinct colors for plotting discrete clusters.
Uses \code{colors15_cheng} as the base palette. If the number of clusters exceeds
the base palette, additional colors are generated using \code{scales::hue_pal()}.
}
\examples{
safeColorPalette(5)
safeColorPalette(20)

}
