% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpNeigh-package.R
\docType{package}
\name{SpNeigh-package}
\alias{SpNeigh}
\alias{SpNeigh-package}
\title{SpNeigh: Spatial Neighborhood Modeling and Differential Expression Analysis for Transcriptomics}
\description{
SpNeigh provides methods for neighborhood-aware analysis of spatial transcriptomics data. It supports boundary detection, spatial weighting (centroid- and boundary-based), spatially informed differential expression using spline-based models, and spatial enrichment analysis via the Spatial Enrichment Index (SEI). Designed for compatibility with Seurat objects, SpatialExperiment objects and spatial data frames, SpNeigh enables interpretable, publication-ready analysis of spatial gene expression patterns.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jinming-cheng/SpNeigh}
  \item Report bugs at \url{https://github.com/jinming-cheng/SpNeigh/issues}
}

}
\author{
\strong{Maintainer}: Jinming Cheng \email{jinming.cheng@outlook.com} (\href{https://orcid.org/0000-0003-3806-4694}{ORCID})

Authors:
\itemize{
  \item Jinming Cheng \email{jinming.cheng@outlook.com} (\href{https://orcid.org/0000-0003-3806-4694}{ORCID})
}

}
\keyword{internal}
