% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{AddBoundary}
\alias{AddBoundary}
\title{Add a boundary or boundaries to the plot}
\usage{
AddBoundary(
  boundary = NULL,
  color_boundary = "black",
  linewidth_boundary = 1.5
)
}
\arguments{
\item{boundary}{Boundary obtained from \code{GetBoundary} function.}

\item{color_boundary}{Color of the boundary for a cluster. Default is black.}

\item{linewidth_boundary}{Linewidth of the boundary borders. Default is 1.5.}
}
\description{
Add a boundary or boundaries to the plot
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Add boundaries to an existing plot
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               subregion_method = "dbscan",
                               eps = 120, minPts = 10)
PlotBoundary(coords) + AddBoundary(boundary_points)

}
