% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{AddBoundaryPoly}
\alias{AddBoundaryPoly}
\title{Add a boundary poly to the plot}
\usage{
AddBoundaryPoly(
  boundary_poly,
  color_boundary = "black",
  linewidth_boundary = 1.5
)
}
\arguments{
\item{boundary_poly}{A sf object returned by \code{BuildBoundaryPoly()}, \code{GetOuterBoundary()} or \code{GetRingRegion()}.}

\item{color_boundary}{Color of the boundary for a cluster. Default is black.}

\item{linewidth_boundary}{Linewidth of the boundary for a cluster. Default is 1.5.}
}
\description{
Add a boundary poly to the plot
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Build boundary polygons from the boundary points
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               subregion_method = "dbscan",
                               eps = 120, minPts = 10)
boundary_polys <- BuildBoundaryPoly(boundary_points)

# Add boundary polygons to an existing plot
PlotBoundary(coords) + AddBoundaryPoly(boundary_polys,
                                       color="blue")

}
