% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{addBoundaryPoly}
\alias{addBoundaryPoly}
\title{Add boundary polygons or linestrings to a spatial plot}
\usage{
addBoundaryPoly(
  boundary_poly,
  color_boundary = "black",
  linewidth_boundary = 1.5
)
}
\arguments{
\item{boundary_poly}{An \code{sf} object containing \code{POLYGON} or \code{LINESTRING}
geometries and a \code{region_id} column.}

\item{color_boundary}{Color for boundary lines. Default is \code{"black"}.}

\item{linewidth_boundary}{Numeric. Line width for boundary outlines.
Default is 1.5.}
}
\value{
A \code{ggplot2::geom_sf} layer that can be added to an existing plot.
}
\description{
Overlays boundary polygons or linestrings
on a spatial \code{ggplot2} plot. This function adds an \code{sf} geometry layer
to display complete boundary shapes for visualizing spatial clusters,
rings, or enriched zones.
}
\examples{
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
    package = "SpNeigh"
))

boundary_points <- getBoundary(
    data = coords, one_cluster = 2,
    subregion_method = "dbscan",
    eps = 120, minPts = 10
)
boundary_polys <- buildBoundaryPoly(boundary_points)

# Add boundary polygons to the plot
plotBoundary(coords) +
    addBoundaryPoly(boundary_polys, color_boundary = "blue")

}
