% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{BoundaryPolyToPoints}
\alias{BoundaryPolyToPoints}
\title{Extract boundary points from boundary polygons}
\usage{
BoundaryPolyToPoints(boundary_poly)
}
\arguments{
\item{boundary_poly}{An \code{sf} object with POLYGON or LINESTRING geometries.}
}
\value{
A \code{data.frame} with columns \code{x}, \code{y} and the specified \code{region_id}.
}
\description{
This function extracts x and y coordinates from a POLYGON or LINESTRING \code{sf} object
by casting it into POINT geometries.
}
\examples{
# Load coordinates
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
                              package = "SpNeigh"))
head(coords)

# Build boundary polygons from the boundary points
boundary_points <- GetBoundary(data = coords, one_cluster = 2)
head(boundary_points)
dim(boundary_points)
boundary_polys <- BuildBoundaryPoly(boundary_points)

# Convert boundary polygons to boundary points
boundary_points2 <- BoundaryPolyToPoints(boundary_polys)
head(boundary_points2)
dim(boundary_points2)

}
