% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{BuildBoundaryPoly}
\alias{BuildBoundaryPoly}
\title{Build boundary polygon based on the boundary points}
\usage{
BuildBoundaryPoly(boundary = NULL)
}
\arguments{
\item{boundary}{A data frame. Boundary result returned by \code{GetBoundary} function.}
}
\description{
Build boundary polygon based on the boundary points
}
\examples{
# Load coordinates
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
                              package = "SpNeigh"))
head(coords)

# Get boundary points of cluster 2
boundary_points <- GetBoundary(data = coords, one_cluster = 2)

# Build boundary polygon from the boundary points
boundary_polys <- BuildBoundaryPoly(boundary_points)

# Plot boundary polygons
plot(boundary_polys)

}
