% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colors15_cheng}
\alias{colors15_cheng}
\title{Custom Color Palette}
\format{
A character vector of length 15.
}
\usage{
colors15_cheng
}
\description{
A character vector containing 15 colors.
}
\examples{
plot(1:15, 1:15, pch = 16, col = colors15_cheng)
barplot(rep(1, 15), col = colors15_cheng)

}
\keyword{datasets}
