% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialEnrichmentIndex.R
\name{ComputeSEI}
\alias{ComputeSEI}
\title{Compute Spatial Enrichment Index (SEI)}
\usage{
ComputeSEI(...)
}
\value{
A data frame containing the spatial enrichment index (SEI) results.
}
\description{
Alias for \code{\link{ComputeSpatialEnrichmentIndex}}.
}
\examples{
# Load spatial coordinates and log-normalized expression
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
    package = "SpNeigh"
))
logNorm_expr <- readRDS(system.file("extdata", "LogNormExpr.rds",
    package = "SpNeigh"
))

# Compute spatial weights and SEI
bon_c0 <- GetBoundary(data = coords, one_cluster = 0)
cells_c0 <- subset(coords, cluster == 0)$cell
weights <- ComputeBoundaryWeights(
    data = coords,
    cell_ids = cells_c0,
    boundary = bon_c0
)
sei_df <- ComputeSEI(logNorm_expr[, cells_c0], weights)
head(sei_df)
}
