% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialInteractionMatrix.R
\name{ComputeSpatialInteractionMatrix}
\alias{ComputeSpatialInteractionMatrix}
\title{Compute a spatial neighborhood interaction matrix based on K-nearest neighbors algorithm}
\usage{
ComputeSpatialInteractionMatrix(data = NULL, k = 10)
}
\arguments{
\item{data}{A data frame with columns: \code{x}, \code{y}, \code{cell}, \code{cluster}. If the input is a Seurat object, the coordinates of cells and cluster information will be extracted automatically.}

\item{k}{Number of nearest neighbors to use. Default is 10.}
}
\value{
A matrix where rows are focal clusters and columns are neighbor clusters.
}
\description{
The spatial neighborhood interaction matrix is computed based on counting the number of clusters of k-nearest neighbor cells by the cluster a focal cell.
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Compute KNN spatial interaction matrix of all cells
interaction_matrix <- ComputeSpatialInteractionMatrix(coords)
interaction_matrix

# Compute KNN spatial interaction matrix of cells insides boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
cells_inside <- GetCellsInside(data = coords, boundary =  boundary_points)
coords_sub <- subset(coords, cell \%in\% cells_inside$cell)
interaction_matrix <- ComputeSpatialInteractionMatrix(coords_sub)
interaction_matrix

# Compute KNN spatial interaction matrix of cells insides rings
ring_regions <- GetRingRegion(boundary = boundary_points, dist = 100)
cells_ring <- GetCellsInside(data = coords, boundary =  ring_regions)
coords_sub <- subset(coords, cell \%in\% cells_ring$cell)
interaction_matrix <- ComputeSpatialInteractionMatrix(coords_sub)
interaction_matrix
}
