% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{ExtractCoords}
\alias{ExtractCoords}
\title{Extract coordinates from a Seurat object or a data frame}
\usage{
ExtractCoords(data = NULL)
}
\arguments{
\item{data}{A data frame contains coordinates of the cells and cluster information. If the input is a Seurat object, the information will be extracted automatically.}
}
\description{
Extract coordinates from a Seurat object or a data frame
}
