% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{ExtractCoords}
\alias{ExtractCoords}
\title{Extract coordinates from a Seurat object or a data frame}
\usage{
ExtractCoords(data = NULL)
}
\arguments{
\item{data}{A data frame with columns: \code{x}, \code{y}, \code{cell}, \code{cluster}. If the input is a Seurat object, the coordinates of cells and cluster information will be extracted automatically.}
}
\description{
Extract coordinates from a Seurat object or a data frame
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))

# An exmaple of the coordinate data.frame used for analysis
head(coords)

# If the data is a data.frame, returns it without doing anything
head(ExtractCoords(coords))

}
