% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotInteractionMatrix.R
\name{FactorNaturalOrder}
\alias{FactorNaturalOrder}
\title{Make a factor of natural order}
\usage{
FactorNaturalOrder(x)
}
\arguments{
\item{x}{A vector}
}
\description{
Make a factor of natural order
}
\examples{
FactorNaturalOrder(10:1)
FactorNaturalOrder( c("a11", "a12", "a1","a2", "a") )

}
