% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{GetBoundary}
\alias{GetBoundary}
\title{Get boundary points for a cluster or cell population}
\usage{
GetBoundary(
  data = NULL,
  one_cluster = NULL,
  k = 5,
  distance_cutoff = 30,
  multi_region = TRUE,
  subregion_method = c("dbscan", "kmeans")[1],
  eps = 80,
  minPts = 10,
  n_subregions = 3
)
}
\arguments{
\item{data}{A data frame contains coordinates of the cells and cluster information. If the input is a Seurat object, the information will be extracted automatically.}

\item{one_cluster}{Specify one cluster to draw the boundary}

\item{k}{The maximum number of nearest neighbors to search, which is a parameter used in \code{get.knn} function. Default is set to 5.}

\item{distance_cutoff}{K-nearest neighbor distance threshold. Default is set to 30. k and distance_cutoff are used to remove outlier points.}

\item{multi_region}{A logical value demonstrates whether a cluster has multiple regions on a spatial plot. Default is TRUE}

\item{subregion_method}{If \code{multi_region} is TRUE, set a method to detect the subregions. Default is "dbscan". The "dbscan" method can be used to automatically detect subregions. Alternatively, "kmeans" method can be used to manually set the number of subregions.}

\item{eps}{A parameter used when \code{subregion_method} is set to "dbscan". Size of the epsilon neighborhood.}

\item{minPts}{A parameter used when \code{subregion_method} is set to "dbscan". Number of minimum points required in the eps neighborhood for core points. The \code{eps} and \code{minPts} parameters are used to adjust the subregion numbers for "dbscan" method.}

\item{n_subregions}{A parameter used when \code{subregion_method} is set to "kmeans". Manually set the number of subregions. Default is 3.}
}
\description{
Get boundary points for a cluster or cell population
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Get boundary points of cluster 1 using dbscan method by default
boundary_points <- GetBoundary(data = coords, one_cluster = 1)

# An example of data in the boundary_points data.frame
head(boundary_points)

# Number of subregions
table(boundary_points$region_id)

# Plot coordinates without boundaries
PlotBoundary(coords)

# Plot boundary points of cluster 1
plot(boundary_points[,c("x","y")], pch=16, cex=0.3)

# Plot boundary polygon of cluster 1
PlotBoundary(coords, one_cluster = 1)

# Get boundary points of cluster 1 using kmeans method
# and manually specify subregion number
boundary_points <- GetBoundary(data = coords, one_cluster = 1,
                               subregion_method = "kmeans",
                               n_subregions = 2)
table(boundary_points$region_id)
plot(boundary_points[,c("x","y")], pch=16, cex=0.3)

# Get boundary points of cluster 1 using concaveman hull method.
# region_id is set to 1 but it is not meaningful for this method.
boundary_points <- GetBoundary(data = coords, one_cluster = 1,
                               multi_region = FALSE)
table(boundary_points$region_id)
plot(boundary_points[,c("x","y")], pch=16, cex=0.3)
}
