% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellsInside.R
\name{GetCellsInside}
\alias{GetCellsInside}
\title{Get cells inside boundaries or rings}
\usage{
GetCellsInside(data = NULL, boundary = NULL)
}
\arguments{
\item{data}{A data frame contains coordinates of the cells and cluster information. If the input is a Seurat object, the information will be extracted automatically.}

\item{boundary}{A sf object or a data frame. Boundary result returned by \code{GetBoundary} or \code{GetBoundaryPoly} functions. Or ring regions returned by \code{GetRingRegion}.}
}
\description{
Get cells inside boundaries or rings
}
