% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{GetOuterBoundary}
\alias{GetOuterBoundary}
\title{Get outer boundary of the original boundary}
\usage{
GetOuterBoundary(boundary = NULL, dist = 100)
}
\arguments{
\item{boundary}{A data frame or sf object. Boundary result returned by \code{GetBoundary} or \code{GetBoundaryPoly} function.}

\item{dist}{Distance of the outer boundary from the existing boundary.}
}
\description{
Get outer boundary of the original boundary
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Get boundary points of cluster 2
boundary_points <- GetBoundary(data = coords, one_cluster = 2)

# Build boundary polygon from the boundary points
boundary_polys = BuildBoundaryPoly(boundary_points)

# Get outer boundary ploygons that are of a distance of 100 units from the original boundaries
# both boundary points and boundary polygons can be used as input
outer_boundary_polys <- GetOuterBoundary(boundary_points, dist = 100)
outer_boundary_polys <- GetOuterBoundary(boundary_polys, dist = 100)

# Plot original boundary polygons
plot(boundary_polys)

# Plot outer boundary polygons
plot(outer_boundary_polys)

}
