% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{GetRingRegion}
\alias{GetRingRegion}
\title{Get ring region between the original boundary and outer boundary}
\usage{
GetRingRegion(boundary = NULL, outer_boundary = NULL, ...)
}
\arguments{
\item{boundary}{A data frame or sf object. Boundary result returned by \code{GetBoundary} or \code{GetBoundaryPoly} function.}

\item{outer_boundary}{A sf object. Outer boundary returned by \code{GetOuterBoundary} function.}

\item{...}{Parameters in \code{GetOuterBoundary}}
}
\description{
Get ring region between the original boundary and outer boundary
}
\examples{
# Load coordinates
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
                              package = "SpNeigh"))
head(coords)

# Get boundary points of cluster 2
boundary_points <- GetBoundary(data = coords, one_cluster = 2)

# Get ring regions (outer regions with be automatically obtained)
ring_regions <- GetRingRegion(boundary = boundary_points, dist = 100)

# Plot ring regions
plot(ring_regions)

# Alternatively, get ring region using the provided boundary and outer_boundary
outer_boundary_polys <- GetOuterBoundary(boundary_points, dist = 100)
ring_regions <- GetRingRegion(boundary = boundary_points,
                              outer_boundary = outer_boundary_polys)
plot(ring_regions)

}
