% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{PlotBoundary}
\alias{PlotBoundary}
\title{Draw boundary for a cluster or cell population}
\usage{
PlotBoundary(
  data = NULL,
  one_cluster = NULL,
  boundary = NULL,
  colors = my_colors_15,
  point_size = 0.5,
  color_boundary = "black",
  linewidth_boundary = 1.5,
  sub_plot = FALSE,
  theme_ggplot = my_theme_ggplot(),
  legend_size = 2,
  ...
)
}
\arguments{
\item{data}{A data frame with columns: \code{x}, \code{y}, \code{cell}, \code{cluster}. If the input is a Seurat object, the coordinates of cells and cluster information will be extracted automatically.}

\item{one_cluster}{Specify one cluster to obtain its boundary. If \code{boundary} parameter is provided, it can be omitted. However, it must be provided, if \code{sub_plot} is TRUE.}

\item{boundary}{Boundary obtained from \code{GetBoundary} function. If it is NULL, get the boundary of the specified cluster automatically. If both \code{boundary} and \code{one_cluster} are NULL, plot coordinates only.}

\item{colors}{A vector of colors used in plotting for clusters.}

\item{point_size}{Point size of cells. Default is 0.5.}

\item{color_boundary}{Color of the boundary for a cluster. Default is black.}

\item{linewidth_boundary}{Linewidth of the boundary for a cluster. Default is 1.5.}

\item{sub_plot}{A logical value indicates whether to only plot cells in one cluster. If FALSE, plot all cells in all clusters. Default is FALSE.}

\item{theme_ggplot}{Theme for ggplot.}

\item{legend_size}{Legend size for ggplot.}

\item{...}{Parameters in \code{GetBoundary} function.}
}
\description{
Draw boundary for a cluster or cell population
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Plot coordinates without boundary
PlotBoundary(coords)

# Plot boundaries of cluster 2
PlotBoundary(coords, one_cluster = 2)

# Plot boundaries of cluster 2, and adjust subregion numbers
# for dbscan method using eps and minPts parameters
PlotBoundary(coords, one_cluster = 2,
            multi_region = TRUE,
            subregion_method = "dbscan",
            eps = 120, minPts = 10)

# Alternatively, the boundaries can be manually specified
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               subregion_method = "dbscan",
                               eps = 120, minPts = 10)
PlotBoundary(data = coords, boundary = boundary_points)
}
