% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCellsInside.R
\name{PlotCellsInside}
\alias{PlotCellsInside}
\title{Plot cells inside boundries or rings}
\usage{
PlotCellsInside(
  cells_inside = NULL,
  point_size = 0.5,
  colors = my_colors_15,
  theme_ggplot = my_theme_ggplot(),
  legend_size = 2,
  ...
)
}
\arguments{
\item{cells_inside}{A sf object. cells_inside returned by \code{GetCellsInside}.}

\item{point_size}{Point size of cells. Default is 0.5.}

\item{colors}{A vector of colors used in plotting for clusters.}

\item{theme_ggplot}{Theme for ggplot.}

\item{legend_size}{Legend size for ggplot.}

\item{...}{Other paramters in \code{sf::geom_sf()} function.}
}
\description{
Plot cells inside boundries or rings
}
