% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCellsInside.R
\name{PlotCellsInside}
\alias{PlotCellsInside}
\title{Plot cells inside boundries or rings}
\usage{
PlotCellsInside(
  cells_inside = NULL,
  point_size = 0.5,
  colors = my_colors_15,
  theme_ggplot = my_theme_ggplot(),
  legend_size = 2,
  fixed_aspect_ratio = TRUE
)
}
\arguments{
\item{cells_inside}{A sf object. \code{cells_inside} is returned by \code{GetCellsInside}.}

\item{point_size}{Point size of cells. Default is 0.5.}

\item{colors}{A character vector specifying the colors for clusters.}

\item{theme_ggplot}{Theme for ggplot.}

\item{legend_size}{Legend size for ggplot.}

\item{fixed_aspect_ratio}{Logical. Whether to plot using \code{sf::geom_sf()} function with 1:1 aspect ratio. Default is TRUE}
}
\description{
Plot cells inside boundries or rings
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Get cells inside boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
cells_inside <- GetCellsInside(data = coords, boundary =  boundary_points)
PlotCellsInside(cells_inside = cells_inside)

# Get cells inside rings
ring_regions <- GetRingRegion(boundary = boundary_points, dist = 100)
cells_ring <- GetCellsInside(data = coords, boundary =  ring_regions)
PlotCellsInside(cells_inside = cells_ring)

}
