% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{plotEdge}
\alias{plotEdge}
\title{Plot boundary edges or segmented boundary lines}
\usage{
plotEdge(
  boundary_poly = NULL,
  linewidth_boundary = 1,
  theme_ggplot = theme_spneigh(),
  ...
)
}
\arguments{
\item{boundary_poly}{An \code{sf} object containing \code{POLYGON} or \code{LINESTRING}
geometries and a \code{region_id} column.}

\item{linewidth_boundary}{Numeric value specifying the line width of
the edge. Default is \code{1}.}

\item{theme_ggplot}{A ggplot2 theme object. Default is \code{theme_spneigh()}.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_sf()}.}
}
\value{
A \code{ggplot} object displaying the edge outlines colored
by \code{region_id}.
}
\description{
Plots boundary edges as colored \code{LINESTRING} or \code{POLYGON} outlines
using \code{ggplot2}.
This function is especially useful for visualizing specific edges
extracted from a polygon using \code{splitBoundaryPolyByAnchor()}.
}
\examples{
coords <- readRDS(system.file("extdata", "MouseBrainCoords.rds",
    package = "SpNeigh"
))

# Build boundary polygon and plot its outline
boundary_points <- getBoundary(
    data = coords, one_cluster = 2,
    eps = 120, minPts = 10
)
boundary_polys <- buildBoundaryPoly(boundary_points)
plotEdge(boundary_poly = boundary_polys)

# Split a polygon into edge segments and plot
boundary_edges <- splitBoundaryPolyByAnchor(boundary_polys[1, ])
plotEdge(boundary_poly = boundary_edges)

}
