% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotExpression.R
\name{PlotExpression}
\alias{PlotExpression}
\title{Plot expression of genes in cells on the spatial plot}
\usage{
PlotExpression(
  data = NULL,
  exp_mat = NULL,
  genes = NULL,
  sub_plot = FALSE,
  one_cluster = NULL,
  split_by = NULL,
  return_list = FALSE,
  point_size = 0.2,
  theme_ggplot = my_theme_ggplot()
)
}
\arguments{
\item{data}{A data frame contains coordinates of the cells and cluster information. If the input is a Seurat object, the information will be extracted automatically.}

\item{exp_mat}{A gene expression matrix or dgCMatrix. If \code{data} is a Seurat object, the \code{exp_mat} will be automatically obtained.}

\item{genes}{Specify genes for plotting the expression.}

\item{sub_plot}{A logical value indicates whether to only plot cells in one cluster. Default is FALSE.}

\item{one_cluster}{Specify a cluster when \code{sub_plot} is TRUE.}

\item{split_by}{A column name in the coordinate data.frame (from \code{data} parameter) used to split the plot.}

\item{return_list}{A logical value indicates whether to return a list of ggplot objects. Default is FALSE.}

\item{point_size}{Point size of cells. Default is 0.2.}

\item{theme_ggplot}{Theme for ggplot.}
}
\description{
Plot expression of genes in cells on the spatial plot
}
\examples{
# Sample data
df <- data.frame(x = c( rnorm(100, mean=1),rnorm(100, mean=5) ),
                 y = c( rnorm(100, mean=1),rnorm(100, mean=5) ),
                 cell = 1:200,
                 cluster = rep(rep(1:2, each = 100)) )
exp_mat <- data.frame(gene1 = c(runif(100, min = 4, max = 6),
                                runif(100, min = 0, max = 2) ),
                      gene2 = c(runif(100, min = 0, max = 1),
                                runif(100, min = 5, max = 8) ) )
exp_mat <- t(exp_mat)
colnames(exp_mat) <- df$cell

# Plot expression of two genes
p <- PlotExpression(data = df, exp_mat = exp_mat,
                    genes = c("gene1","gene2"),point_size = 2)
p

# Plot expression of gene1 in cluster 1
p <- PlotExpression(data = df, exp_mat = exp_mat,
                    sub_plot = TRUE, one_cluster = 1,
                    genes = c("gene1"),point_size = 2)
p + ggplot2::facet_wrap(~cluster) +
  ggplot2::scale_fill_viridis_c(option = "plasma",limits = c(0, 6)) +
  ggplot2::scale_color_viridis_c(option = "plasma",limits = c(0, 6))

}
