% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotInteractionMatrix.R
\name{PlotInteractionMatrix}
\alias{PlotInteractionMatrix}
\title{Heatmap of row-scaled interaction matrix}
\usage{
PlotInteractionMatrix(
  interaction_matrix,
  low_color = "blue",
  mid_color = "white",
  high_color = "red",
  angle_x_label = 45,
  title = "Row-Scaled Interaction Matrix (Z-scores)"
)
}
\arguments{
\item{interaction_matrix}{A numeric matrix where rows represent focal clusters and columns represent neighbor clusters. For example, the output \code{interaction_matrix} from \code{ComputeSpatialInteractionMatrix}.}

\item{low_color}{Color for low values. Default is blue.}

\item{mid_color}{Color for mid-range values. Default is white.}

\item{high_color}{Color for high values. Default: red.}

\item{angle_x_label}{Angle to rotate the x-axis labels.}

\item{title}{Title for the heatmap.}
}
\description{
Heatmap of row-scaled interaction matrix
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)


# Compute KNN spatial interaction matrix of cells insides boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
ring_regions <- GetRingRegion(boundary = boundary_points, dist = 100)
cells_ring <- GetCellsInside(data = coords, boundary =  ring_regions)
coords_sub <- subset(coords, cell \%in\% cells_ring$cell)
interaction_matrix <- ComputeSpatialInteractionMatrix(coords_sub)

# Plot interaction matrix
PlotInteractionMatrix(interaction_matrix)
}
