% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotBoundary.R
\name{PlotRegion}
\alias{PlotRegion}
\title{Plot regions inside boundries or rings}
\usage{
PlotRegion(
  boundary_poly = NULL,
  alpha = 0.5,
  color_boundary = "black",
  linewidth_boundary = 1,
  theme_ggplot = my_theme_ggplot(),
  ...
)
}
\arguments{
\item{boundary_poly}{A sf object returned by \code{BuildBoundaryPoly()}, \code{GetOuterBoundary()} or \code{GetRingRegion()}.}

\item{alpha}{The transparency for the fill aesthetics of a geom in ggplot.}

\item{color_boundary}{Color of the boundaries. Default is black.}

\item{linewidth_boundary}{Linewidth of the boundaries. Default is 1.}

\item{theme_ggplot}{Theme for ggplot.}

\item{...}{Other paramters in \code{sf::geom_sf()} function.}
}
\description{
Plot regions inside boundries or rings
}
