% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStats.R
\name{PlotStatsBar}
\alias{PlotStatsBar}
\title{Bar plot of statistics of cells inside boundaries or rings}
\usage{
PlotStatsBar(
  cell_stats = NULL,
  stat_column = c("proportion", "count")[1],
  colors = my_colors_15,
  angle_x_label = 0,
  theme_ggplot = my_theme_ggplot()
)
}
\arguments{
\item{cell_stats}{Statistics of cells returned by \code{StatsCellsInside()}.}

\item{stat_column}{A column name in \code{cell_stats} and used for y axis.}

\item{colors}{A vector of colors used in plotting for clusters.}

\item{angle_x_label}{Angle to rotate the x-axis labels.}

\item{theme_ggplot}{Theme for ggplot.}
}
\description{
Bar plot of statistics of cells inside boundaries or rings
}
