% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStats.R
\name{PlotStatsBar}
\alias{PlotStatsBar}
\title{Bar plot of statistics of cells inside boundaries or rings}
\usage{
PlotStatsBar(
  cell_stats = NULL,
  stat_column = c("proportion", "count"),
  colors = my_colors_15,
  angle_x_label = 0,
  theme_ggplot = my_theme_ggplot()
)
}
\arguments{
\item{cell_stats}{Statistics of cells returned by \code{StatsCellsInside()}.}

\item{stat_column}{A column name in \code{cell_stats} and used for y axis.}

\item{colors}{A character vector specifying the colors for clusters.}

\item{angle_x_label}{Angle to rotate the x-axis labels.}

\item{theme_ggplot}{Theme for ggplot.}
}
\description{
Bar plot of statistics of cells inside boundaries or rings
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Obtain statistics of cells inside boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
cells_inside <- GetCellsInside(data = coords, boundary =  boundary_points)
stats_cells = StatsCellsInside(cells_inside)

# Plot proportion of cells in different clusters using bar plot
PlotStatsBar(stats_cells, stat_column = "proportion")

# Plot count of cells in different clusters using bar plot
PlotStatsBar(stats_cells, stat_column = "count")

}
