% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStats.R
\name{PlotStatsPie}
\alias{PlotStatsPie}
\title{Pie chart or donut chart of statistics of cells inside boundaries or rings}
\usage{
PlotStatsPie(
  cell_stats = NULL,
  plot_donut = FALSE,
  add_labels = TRUE,
  label_cutoff = 0.01,
  label_color = "white",
  label_size = 4,
  label_nudge_x = 0.1,
  colors = my_colors_15
)
}
\arguments{
\item{cell_stats}{Statistics of cells returned by \code{StatsCellsInside()}.}

\item{plot_donut}{A logical value indicates whether to make a donut chart. Default is FALSE.}

\item{add_labels}{A logical value indicates whether to add percentage labels inside the plot. Default is TRUE.}

\item{label_cutoff}{Labels of proportion smaller than this cutoff will not be plotted. Default is 0.01 (or 1\%). Set it to 0 to plot labels of all percentages.}

\item{label_color}{Color of the percentage labels.}

\item{label_size}{Text size of the percentage labels.}

\item{label_nudge_x}{Nudge x coordinate to adjust positions for the percentage labels.}

\item{colors}{A character vector specifying the colors for clusters.}
}
\description{
Pie chart or donut chart of statistics of cells inside boundaries or rings
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Obtain statistics of cells inside boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
cells_inside <- GetCellsInside(data = coords, boundary =  boundary_points)
stats_cells = StatsCellsInside(cells_inside)

# Plot proportion using pie chart without adding percentage labels
PlotStatsPie(stats_cells, add_labels = FALSE)

# Plot proportion using pie chart showing labels of all percentages
PlotStatsPie(stats_cells, label_cutoff = 0)

# Plot proportion using pie chart showing labels of percentages greater than 1\%
PlotStatsPie(stats_cells, label_cutoff = 0.01)

# Plot proportion using donut chart
PlotStatsPie(stats_cells, plot_donut = TRUE)

}
