% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWeights.R
\name{PlotWeights}
\alias{PlotWeights}
\title{Plot spatial weights for cells}
\usage{
PlotWeights(
  data = NULL,
  weights = NULL,
  point_size = 0.2,
  theme_ggplot = my_theme_ggplot()
)
}
\arguments{
\item{data}{A data frame or Seurat object containng at least three columns: \code{x}, \code{y} and \code{cell}.}

\item{weights}{A named numeric vector where names are cell IDs.}

\item{point_size}{Point size in the plot. Default is 0.2.}

\item{theme_ggplot}{Theme for ggplot.}
}
\description{
Plots spatial coordinates of cells colored by weight values. Only cells in the \code{data} with weights will be plotted.
}
