% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBoundary.R
\name{RemoveOutliers}
\alias{RemoveOutliers}
\title{Removes outlier cells that are too far from their neighbors in a cluster or cell population}
\usage{
RemoveOutliers(coords, k = 5, distance_cutoff = 30)
}
\arguments{
\item{coords}{coordinates of the cells}

\item{k}{The maximum number of nearest neighbors to search, which is a parameter used in \code{get.knn} function. Default is set to 5.}

\item{distance_cutoff}{K-nearest neighbor distance Threshold. Default is set to 30. k and distance_cutoff are used to remove outlier points.}
}
\description{
Removes outlier cells that are too far from their neighbors in a cluster or cell population
}
