% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SafeColorPalette}
\alias{SafeColorPalette}
\title{Generate a safe color palette for discrete clusters}
\usage{
SafeColorPalette(n_clusters, base_colors = my_colors_15, verbose = TRUE)
}
\arguments{
\item{n_clusters}{Number of unique clusters.}

\item{base_colors}{A character vector of base colors. Default is \code{my_colors_15}.}

\item{verbose}{Logical. If \code{TRUE}, displays a message when extended colors are used. Default is \code{TRUE}.}
}
\value{
A character vector of colors of length \code{n_clusters}.
}
\description{
Returns a vector of visually distinct colors for plotting discrete clusters.
Uses \code{my_colors_15} as the base palette. If the number of clusters exceeds
the base palette, additional colors are generated using \code{scales::hue_pal()}.
}
\examples{
SafeColorPalette(5)
SafeColorPalette(20)

}
