% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsCellsInside.R
\name{StatsCellsInside}
\alias{StatsCellsInside}
\title{Statistics of cells inside boundaries or rings}
\usage{
StatsCellsInside(cells_inside = NULL)
}
\arguments{
\item{cells_inside}{A sf object. cells_inside returned by \code{GetCellsInside}.}
}
\description{
Statistics of cells inside boundaries or rings
}
