% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatsCellsInside.R
\name{StatsCellsInside}
\alias{StatsCellsInside}
\title{Statistics of cells inside boundaries or rings}
\usage{
StatsCellsInside(cells_inside = NULL)
}
\arguments{
\item{cells_inside}{A sf object. cells_inside returned by \code{GetCellsInside}.}
}
\description{
Statistics of cells inside boundaries or rings
}
\examples{
# Load coordinates
load(system.file("extdata", "MouseBrainTinyCoords.rda",
                 package = "SpNeigh"))
head(coords)

# Get cells inside boundaries
boundary_points <- GetBoundary(data = coords, one_cluster = 2,
                               eps = 120, minPts = 10)
cells_inside <- GetCellsInside(data = coords, boundary =  boundary_points)

# Statistics of cells inside boundaries
stats_cells = StatsCellsInside(cells_inside)
head(stats_cells)

}
